+++ src/dune_rules/workspace.ml
@@ -662,7 +662,10 @@ let workspace =
   Memo.exec memo
 
 let update_execution_parameters t ep =
-  Execution_parameters.set_swallow_stdout_on_success
-    t.config.swallow_stdout_on_success ep
+  ep
+  |> Execution_parameters.set_action_stdout_on_success
+       t.config.action_stdout_on_success
+  |> Execution_parameters.set_action_stderr_on_success
+       t.config.action_stderr_on_success
 
 let build_contexts t = List.concat_map t.contexts ~f:Context.build_contexts